minor change.
git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@10438 8b29bb6e-589a-435a-9a12-3ab907be3462

+++ components/interactions/src/org/geworkbench/components/interactions/cellularnetwork/ThrottleGraph.java
@@ -239,7 +239,7 @@ public class ThrottleGraph extends JPanel {
 			double b = maxConfidenceValue / (Math.pow(10, a));
 			double maxX = Math.round(b);
 			maxX = maxX * (Math.pow(10, a));
-			long smallestIncrement = (long) maxX / 100;
+			double smallestIncrement = maxX / 100;
 
 			tgPreference
 					.setSmallestIncrement(smallestIncrement);
